Снимает обработчик события и удаляет ненужные функции
